feature(#403): NURBS×Primitives 統合と分類軸統一（nurbs_3d化）
- collision/nurbs_3d.rs: NurbsCurveCollider + NurbsSurfaceCollider の統合実装
- intersection/nurbs_3d.rs: 両方の交点計算関数を統合
- distance/nurbs_3d.rs: NurbsCurve3D 距離関数の拡張（fallible API対応）
- distance/primitive_3d.rs: NURBS関数の移動
- mod.rs: 各モジュール宣言を再構成
- coverage_matrix_engine.rs: 20個のMatrixEntry ID を nurbs_3d 統一基準へ更新
- GEO_ALGORITHMS_MODULE_STRUCTURE_RULES.md: 分割ルール正規化ドキュメント
- ISSUE_403_IMPLEMENTATION_PREP.md: 実装手順と完了チェック
- 旧ファイル削除: 4個（primitive_nurbs*.rs）

実装ペア数:
- NurbsCurve3D × 9種 Primitive
- NurbsSurface3D × 9種 Primitive
計 17 ペア

統計:
- collision: 18個の distance 関数
- intersection: 18個の intersection 関数
- distance: 27個の関数（正規+逆+try）
- テスト: 85個

分類軸統一:
Before: nurbs_curve_3d / nurbs_surface_3d（形状タイプベース）
After: nurbs_3d（次元ベース、future-proof）
