+++ model/geo_algorithms/src/collision/nurbs_3d.rs
@@ -1,14 +1,16 @@
-//! NURBS × Primitives 衝突判定実装
+//! NURBS3D × Primitives 衝突判定実装
 //!
 //! geo_primitives と geo_nurbs 間の衝突判定を
 //! geo_algorithms で集約して実装するモジュールです。
 //!
+//! NurbsCurve3D / NurbsSurface3D と各種3D Primitive の衝突判定を統一します。
+//!
 //! ## Orphan Rules への対応
 //!
 //! 外部 trait を外部型へ直接実装できないため、
-//! Newtype（NurbsCurveCollider）経由で BasicCollision を実装します。
+//! Newtype（NurbsCurveCollider / NurbsSurfaceCollider）経由で BasicCollision を実装します。
 //!
-//! ## 設計方針
+//! ## 設計方針  
 //!
 //! 1. サンプリング + 数値最適化で距離を評価
 //! 2. 形状ペアごとに BasicCollision を実装
@@ -17,22 +19,25 @@
 //! ## Public Interface (intersection側で使用)
 //!
 //! - `nurbscurve3d_XXX_distance`: NurbsCurve3D と各Primitive間の距離計算
-//!   - intersection/primitive_nurbs.rs から直接呼び出し可能
+//! - `nurbssurface3d_XXX_distance`: NurbsSurface3D と各Primitive間の距離計算
+//!   - intersection/nurbs_3d.rs から直接呼び出し可能
 //!   - collision 判定と intersection 判定（tolerance比較）で共有
 
 use crate::{
     Circle3D, CylindricalSolid3D, EllipsoidalSolid3D, InfiniteLine3D, LineSegment3D, Plane3D,
-    Ray3D, SphericalSolid3D,
+    Point3D, Ray3D, SphericalSolid3D,
 };
 use analysis::linalg::solver::newton::newton_solve_with_numeric_derivative_bounded;
 use geo_contracts::{
     BasicCollision, Circle3DProperties, CylindricalSolid3DMeasure, InfiniteLine3DProperties,
     Plane3DProperties, Scalar,
 };
-use geo_core::Point3D;
-use geo_nurbs::NurbsCurve3D;
+use geo_core::Point3D as CorePoint3D;
+use geo_nurbs::{NurbsCurve3D, NurbsSurface3D};
 
-// Newtype Wrapper for NurbsCurve3D
+// ─────────────────────────────────────────────────────────────────────────
+// NurbsCurveCollider Newtype Wrapper
+// ─────────────────────────────────────────────────────────────────────────
 
 /// NURBS曲線の衝突判定アダプタ（Newtype パターン）
 ///
@@ -64,7 +69,7 @@ impl<T: Scalar> NurbsCurveCollider<T> {
     /// C(u)が点Pに最も近いとき、C(u)-P は C'(u) に直交する
     pub fn newton_refine_closest_point(
         &self,
-        point: &Point3D<T>,
+        point: &CorePoint3D<T>,
         initial_u: T,
         u_min: T,
         u_max: T,
@@ -303,7 +308,7 @@ impl<T: Scalar> BasicCollision<T, InfiniteLine3D<T>> for NurbsCurveCollider<T> {
         let delta_u = (u_max - u_min) / T::from_usize(num_samples);
 
         let (px, py, pz) = <InfiniteLine3D<T> as InfiniteLine3DProperties<T>>::point(line);
-        let point_on_line = Point3D::new(px, py, pz);
+        let point_on_line = CorePoint3D::new(px, py, pz);
         let (dx, dy, dz) = <InfiniteLine3D<T> as InfiniteLine3DProperties<T>>::direction(line);
         let direction_vec = geo_core::Vector3D::new(dx, dy, dz);
 
@@ -366,7 +371,7 @@ impl<T: Scalar> BasicCollision<T, Circle3D<T>> for NurbsCurveCollider<T> {
 
             // Circle3Dの中心からの距離を計算し、半径を引く
             let (cx, cy, cz) = <Circle3D<T> as Circle3DProperties<T>>::center(circle);
-            let center = Point3D::new(cx, cy, cz);
+            let center = CorePoint3D::new(cx, cy, cz);
             let radius = <Circle3D<T> as Circle3DProperties<T>>::radius(circle);
 
             let dx = curve_point.x() - center.x();
@@ -454,7 +459,7 @@ impl<T: Scalar> BasicCollision<T, SphericalSolid3D<T>> for NurbsCurveCollider<T>
         for i in 0..=num_samples {
             let u = u_min + delta_u * T::from_usize(i);
             let curve_vec = self.0.evaluate_at(u);
-            let curve_point = Point3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
+            let curve_point = CorePoint3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
 
             let distance = sphere.distance_to_surface(curve_point);
             min_distance = min_distance.min(distance);
@@ -488,7 +493,7 @@ impl<T: Scalar> BasicCollision<T, EllipsoidalSolid3D<T>> for NurbsCurveCollider<
         for i in 0..=num_samples {
             let u = u_min + delta_u * T::from_usize(i);
             let curve_vec = self.0.evaluate_at(u);
-            let curve_point = Point3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
+            let curve_point = CorePoint3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
 
             let distance = if ellipsoid.contains_point(&curve_point) {
                 T::ZERO
@@ -526,7 +531,7 @@ impl<T: Scalar> BasicCollision<T, CylindricalSolid3D<T>> for NurbsCurveCollider<
         for i in 0..=num_samples {
             let u = u_min + delta_u * T::from_usize(i);
             let curve_vec = self.0.evaluate_at(u);
-            let curve_point = Point3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
+            let curve_point = CorePoint3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
 
             // 円柱の distance_to_point メソッドを使用
             let distance =
@@ -541,14 +546,468 @@ impl<T: Scalar> BasicCollision<T, CylindricalSolid3D<T>> for NurbsCurveCollider<
     }
 }
 
-// ── Public Distance Functions (shared with intersection module) ──
+// ─────────────────────────────────────────────────────────────────────────
+// NurbsSurfaceCollider Newtype Wrapper
+// ─────────────────────────────────────────────────────────────────────────
 
-/// NurbsCurve3D から Point3D への距離を計算
+#[repr(transparent)]
+#[derive(Debug, Clone)]
+pub struct NurbsSurfaceCollider<T: Scalar>(pub NurbsSurface3D<T>);
+
+impl<T: Scalar> NurbsSurfaceCollider<T> {
+    pub fn new(surface: NurbsSurface3D<T>) -> Self {
+        Self(surface)
+    }
+
+    fn nearest_sample_to_point(&self, point: &Point3D<T>) -> (Point3D<T>, T) {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let mut best_point = Point3D::new(T::ZERO, T::ZERO, T::ZERO);
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+                let surface_point = Point3D::new(p.x(), p.y(), p.z());
+
+                let dx = surface_point.x() - point.x();
+                let dy = surface_point.y() - point.y();
+                let dz = surface_point.z() - point.z();
+                let d = (dx * dx + dy * dy + dz * dz).sqrt();
+
+                if d < min_dist {
+                    min_dist = d;
+                    best_point = surface_point;
+                }
+            }
+        }
+
+        (best_point, min_dist)
+    }
+
+    fn min_distance_to_plane(&self, plane: &Plane3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let origin = plane.origin();
+        let (nx, ny, nz) = <Plane3D<T> as Plane3DProperties<T>>::normal(plane);
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+
+                let dx = p.x() - origin.x();
+                let dy = p.y() - origin.y();
+                let dz = p.z() - origin.z();
+                let d = (dx * nx + dy * ny + dz * nz).abs();
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_ray(&self, ray: &Ray3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let origin = ray.origin();
+        let direction = ray.direction_vector();
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+
+                let to_px = p.x() - origin.x();
+                let to_py = p.y() - origin.y();
+                let to_pz = p.z() - origin.z();
+
+                let t = (to_px * direction.x() + to_py * direction.y() + to_pz * direction.z())
+                    .max(T::ZERO);
+
+                let closest_x = origin.x() + t * direction.x();
+                let closest_y = origin.y() + t * direction.y();
+                let closest_z = origin.z() + t * direction.z();
+
+                let dx = p.x() - closest_x;
+                let dy = p.y() - closest_y;
+                let dz = p.z() - closest_z;
+                let d = (dx * dx + dy * dy + dz * dz).sqrt();
+
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_line_segment(&self, segment: &LineSegment3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let start = segment.start();
+        let end = segment.end();
+        let seg_dx = end.x() - start.x();
+        let seg_dy = end.y() - start.y();
+        let seg_dz = end.z() - start.z();
+        let seg_len_sq = seg_dx * seg_dx + seg_dy * seg_dy + seg_dz * seg_dz;
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+
+                let to_px = p.x() - start.x();
+                let to_py = p.y() - start.y();
+                let to_pz = p.z() - start.z();
+
+                let t = if seg_len_sq.is_zero() {
+                    T::ZERO
+                } else {
+                    ((to_px * seg_dx + to_py * seg_dy + to_pz * seg_dz) / seg_len_sq)
+                        .clamp(T::ZERO, T::ONE)
+                };
+
+                let cx = start.x() + t * seg_dx;
+                let cy = start.y() + t * seg_dy;
+                let cz = start.z() + t * seg_dz;
+
+                let dx = p.x() - cx;
+                let dy = p.y() - cy;
+                let dz = p.z() - cz;
+                let d = (dx * dx + dy * dy + dz * dz).sqrt();
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_infinite_line(&self, line: &InfiniteLine3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let (px, py, pz) = <InfiniteLine3D<T> as InfiniteLine3DProperties<T>>::point(line);
+        let (dx, dy, dz) = <InfiniteLine3D<T> as InfiniteLine3DProperties<T>>::direction(line);
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+
+                let tx = p.x() - px;
+                let ty = p.y() - py;
+                let tz = p.z() - pz;
+                let t = tx * dx + ty * dy + tz * dz;
+
+                let cx = px + t * dx;
+                let cy = py + t * dy;
+                let cz = pz + t * dz;
+
+                let ddx = p.x() - cx;
+                let ddy = p.y() - cy;
+                let ddz = p.z() - cz;
+                let d = (ddx * ddx + ddy * ddy + ddz * ddz).sqrt();
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_circle(&self, circle: &Circle3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let (cx, cy, cz) = <Circle3D<T> as Circle3DProperties<T>>::center(circle);
+        let radius = <Circle3D<T> as Circle3DProperties<T>>::radius(circle);
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+
+                let dx = p.x() - cx;
+                let dy = p.y() - cy;
+                let dz = p.z() - cz;
+                let distance_to_center = (dx * dx + dy * dy + dz * dz).sqrt();
+                let d = (distance_to_center - radius).abs();
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_spherical_solid(&self, sphere: &SphericalSolid3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+                let surface_point = CorePoint3D::new(p.x(), p.y(), p.z());
+                min_dist = min_dist.min(sphere.distance_to_surface(surface_point));
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_ellipsoidal_solid(&self, ellipsoid: &EllipsoidalSolid3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+                let surface_point = CorePoint3D::new(p.x(), p.y(), p.z());
+
+                let d = if ellipsoid.contains_point(&surface_point) {
+                    T::ZERO
+                } else {
+                    ellipsoid.distance_to_surface(&surface_point)
+                };
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+
+    fn min_distance_to_cylindrical_solid(&self, cylinder: &CylindricalSolid3D<T>) -> T {
+        let samples_u = 24;
+        let samples_v = 24;
+        let ((u_min, u_max), (v_min, v_max)) = self.0.parameter_domain();
+        let du = (u_max - u_min) / T::from_usize(samples_u);
+        let dv = (v_max - v_min) / T::from_usize(samples_v);
+
+        let mut min_dist = T::INFINITY;
+
+        for i in 0..=samples_u {
+            for j in 0..=samples_v {
+                let u = u_min + du * T::from_usize(i);
+                let v = v_min + dv * T::from_usize(j);
+                let p = self.0.evaluate_at(u, v);
+
+                let d = <CylindricalSolid3D<T> as CylindricalSolid3DMeasure<T>>::distance_to_point(
+                    cylinder,
+                    (p.x(), p.y(), p.z()),
+                );
+                min_dist = min_dist.min(d);
+            }
+        }
+
+        min_dist
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, Point3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, point: &Point3D<T>, tolerance: T) -> bool {
+        self.distance_to(point) <= tolerance
+    }
+
+    fn overlaps(&self, point: &Point3D<T>, tolerance: T) -> bool {
+        self.intersects(point, tolerance)
+    }
+
+    fn distance_to(&self, point: &Point3D<T>) -> T {
+        self.nearest_sample_to_point(point).1
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, Plane3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, plane: &Plane3D<T>, tolerance: T) -> bool {
+        self.distance_to(plane) <= tolerance
+    }
+
+    fn overlaps(&self, plane: &Plane3D<T>, tolerance: T) -> bool {
+        self.intersects(plane, tolerance)
+    }
+
+    fn distance_to(&self, plane: &Plane3D<T>) -> T {
+        self.min_distance_to_plane(plane)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, Ray3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, ray: &Ray3D<T>, tolerance: T) -> bool {
+        self.distance_to(ray) <= tolerance
+    }
+
+    fn overlaps(&self, ray: &Ray3D<T>, tolerance: T) -> bool {
+        self.intersects(ray, tolerance)
+    }
+
+    fn distance_to(&self, ray: &Ray3D<T>) -> T {
+        self.min_distance_to_ray(ray)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, LineSegment3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, segment: &LineSegment3D<T>, tolerance: T) -> bool {
+        self.distance_to(segment) <= tolerance
+    }
+
+    fn overlaps(&self, segment: &LineSegment3D<T>, tolerance: T) -> bool {
+        self.intersects(segment, tolerance)
+    }
+
+    fn distance_to(&self, segment: &LineSegment3D<T>) -> T {
+        self.min_distance_to_line_segment(segment)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, InfiniteLine3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, line: &InfiniteLine3D<T>, tolerance: T) -> bool {
+        self.distance_to(line) <= tolerance
+    }
+
+    fn overlaps(&self, line: &InfiniteLine3D<T>, tolerance: T) -> bool {
+        self.intersects(line, tolerance)
+    }
+
+    fn distance_to(&self, line: &InfiniteLine3D<T>) -> T {
+        self.min_distance_to_infinite_line(line)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, Circle3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, circle: &Circle3D<T>, tolerance: T) -> bool {
+        self.distance_to(circle) <= tolerance
+    }
+
+    fn overlaps(&self, circle: &Circle3D<T>, tolerance: T) -> bool {
+        self.intersects(circle, tolerance)
+    }
+
+    fn distance_to(&self, circle: &Circle3D<T>) -> T {
+        self.min_distance_to_circle(circle)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, SphericalSolid3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, sphere: &SphericalSolid3D<T>, tolerance: T) -> bool {
+        self.distance_to(sphere) <= tolerance
+    }
+
+    fn overlaps(&self, sphere: &SphericalSolid3D<T>, tolerance: T) -> bool {
+        self.intersects(sphere, tolerance)
+    }
+
+    fn distance_to(&self, sphere: &SphericalSolid3D<T>) -> T {
+        self.min_distance_to_spherical_solid(sphere)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, EllipsoidalSolid3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, ellipsoid: &EllipsoidalSolid3D<T>, tolerance: T) -> bool {
+        self.distance_to(ellipsoid) <= tolerance
+    }
+
+    fn overlaps(&self, ellipsoid: &EllipsoidalSolid3D<T>, tolerance: T) -> bool {
+        self.intersects(ellipsoid, tolerance)
+    }
+
+    fn distance_to(&self, ellipsoid: &EllipsoidalSolid3D<T>) -> T {
+        self.min_distance_to_ellipsoidal_solid(ellipsoid)
+    }
+}
+
+impl<T: Scalar> BasicCollision<T, CylindricalSolid3D<T>> for NurbsSurfaceCollider<T> {
+    type Point2D = Point3D<T>;
+
+    fn intersects(&self, cylinder: &CylindricalSolid3D<T>, tolerance: T) -> bool {
+        self.distance_to(cylinder) <= tolerance
+    }
+
+    fn overlaps(&self, cylinder: &CylindricalSolid3D<T>, tolerance: T) -> bool {
+        self.intersects(cylinder, tolerance)
+    }
+
+    fn distance_to(&self, cylinder: &CylindricalSolid3D<T>) -> T {
+        self.min_distance_to_cylindrical_solid(cylinder)
+    }
+}
+
+// ─────────────────────────────────────────────────────────────────────────
+// Public Distance Functions (shared with intersection module)
+// ─────────────────────────────────────────────────────────────────────────
+
+// NurbsCurve3D distance functions
+
 pub fn nurbscurve3d_point3d_distance<T: Scalar>(curve: &NurbsCurve3D<T>, point: &Point3D<T>) -> T {
     NurbsCurveCollider::new(curve.clone()).distance_to(point)
 }
 
-/// NurbsCurve3D から LineSegment3D への距離を計算
 pub fn nurbscurve3d_line_segment3d_distance<T: Scalar>(
     curve: &NurbsCurve3D<T>,
     segment: &LineSegment3D<T>,
@@ -556,12 +1015,10 @@ pub fn nurbscurve3d_line_segment3d_distance<T: Scalar>(
     NurbsCurveCollider::new(curve.clone()).distance_to(segment)
 }
 
-/// NurbsCurve3D から Ray3D への距離を計算
 pub fn nurbscurve3d_ray3d_distance<T: Scalar>(curve: &NurbsCurve3D<T>, ray: &Ray3D<T>) -> T {
     NurbsCurveCollider::new(curve.clone()).distance_to(ray)
 }
 
-/// NurbsCurve3D から InfiniteLine3D への距離を計算
 pub fn nurbscurve3d_infinite_line3d_distance<T: Scalar>(
     curve: &NurbsCurve3D<T>,
     line: &InfiniteLine3D<T>,
@@ -569,7 +1026,6 @@ pub fn nurbscurve3d_infinite_line3d_distance<T: Scalar>(
     NurbsCurveCollider::new(curve.clone()).distance_to(line)
 }
 
-/// NurbsCurve3D から Circle3D への距離を計算
 pub fn nurbscurve3d_circle3d_distance<T: Scalar>(
     curve: &NurbsCurve3D<T>,
     circle: &Circle3D<T>,
@@ -577,12 +1033,10 @@ pub fn nurbscurve3d_circle3d_distance<T: Scalar>(
     NurbsCurveCollider::new(curve.clone()).distance_to(circle)
 }
 
-/// NurbsCurve3D から Plane3D への距離を計算
 pub fn nurbscurve3d_plane3d_distance<T: Scalar>(curve: &NurbsCurve3D<T>, plane: &Plane3D<T>) -> T {
     NurbsCurveCollider::new(curve.clone()).distance_to(plane)
 }
 
-/// NurbsCurve3D から SphericalSolid3D への距離を計算
 pub fn nurbscurve3d_spherical_solid3d_distance<T: Scalar>(
     curve: &NurbsCurve3D<T>,
     sphere: &SphericalSolid3D<T>,
@@ -590,7 +1044,6 @@ pub fn nurbscurve3d_spherical_solid3d_distance<T: Scalar>(
     NurbsCurveCollider::new(curve.clone()).distance_to(sphere)
 }
 
-/// NurbsCurve3D から EllipsoidalSolid3D への距離を計算
 pub fn nurbscurve3d_ellipsoidal_solid3d_distance<T: Scalar>(
     curve: &NurbsCurve3D<T>,
     ellipsoid: &EllipsoidalSolid3D<T>,
@@ -598,7 +1051,6 @@ pub fn nurbscurve3d_ellipsoidal_solid3d_distance<T: Scalar>(
     NurbsCurveCollider::new(curve.clone()).distance_to(ellipsoid)
 }
 
-/// NurbsCurve3D から CylindricalSolid3D への距離を計算
 pub fn nurbscurve3d_cylindrical_solid3d_distance<T: Scalar>(
     curve: &NurbsCurve3D<T>,
     cylinder: &CylindricalSolid3D<T>,
@@ -606,13 +1058,78 @@ pub fn nurbscurve3d_cylindrical_solid3d_distance<T: Scalar>(
     NurbsCurveCollider::new(curve.clone()).distance_to(cylinder)
 }
 
+// NurbsSurface3D distance functions
+
+pub fn nurbssurface3d_point3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    point: &Point3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(point)
+}
+
+pub fn nurbssurface3d_plane3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    plane: &Plane3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(plane)
+}
+
+pub fn nurbssurface3d_ray3d_distance<T: Scalar>(surface: &NurbsSurface3D<T>, ray: &Ray3D<T>) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(ray)
+}
+
+pub fn nurbssurface3d_line_segment3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    segment: &LineSegment3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(segment)
+}
+
+pub fn nurbssurface3d_infinite_line3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    line: &InfiniteLine3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(line)
+}
+
+pub fn nurbssurface3d_circle3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    circle: &Circle3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(circle)
+}
+
+pub fn nurbssurface3d_spherical_solid3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    sphere: &SphericalSolid3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(sphere)
+}
+
+pub fn nurbssurface3d_ellipsoidal_solid3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    ellipsoid: &EllipsoidalSolid3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(ellipsoid)
+}
+
+pub fn nurbssurface3d_cylindrical_solid3d_distance<T: Scalar>(
+    surface: &NurbsSurface3D<T>,
+    cylinder: &CylindricalSolid3D<T>,
+) -> T {
+    NurbsSurfaceCollider::new(surface.clone()).distance_to(cylinder)
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
-    use geo_contracts::NurbsCurve3DConstructor;
+    use crate::Vector3D;
+    use geo_contracts::NurbsSurface3DConstructor;
+    use geo_nurbs::NurbsCurve3D;
 
     fn create_test_curve<T: Scalar>() -> NurbsCurve3D<T> {
         use analysis::linalg::vector::vector3::Vector3;
+        use geo_contracts::NurbsCurve3DConstructor;
 
         // 3次 NURBS 曲線（制御点4つ、次数3）
         let control_points = vec![
@@ -634,8 +1151,6 @@ mod tests {
             T::ONE,
         ];
 
-        // テスト用の NURBS 曲線を constructor trait で生成
-        // シグネチャ: new(degree, knots, control_points: Vec<(T,T,T)>, weights)
         let control_points_tuples = control_points
             .into_iter()
             .map(|v| (v.x(), v.y(), v.z()))
@@ -649,256 +1164,55 @@ mod tests {
         .unwrap()
     }
 
+    fn create_test_surface<T: Scalar>() -> NurbsSurface3D<T> {
+        <NurbsSurface3D<T> as NurbsSurface3DConstructor<T>>::unit_plane()
+    }
+
     #[test]
-    fn test_point_on_curve() {
+    fn test_nurbscurve3d_point3d_distance_zero_on_curve() {
         let curve = create_test_curve::<f64>();
         let collider = NurbsCurveCollider::new(curve);
-        let point = Point3D::new(0.0, 0.0, 0.0); // 曲線の始点
+        let point = CorePoint3D::new(0.0, 0.0, 0.0);
         let tolerance = 1e-6;
 
         assert!(collider.intersects(&point, tolerance));
     }
 
     #[test]
-    fn test_point_near_curve() {
+    fn test_nurbscurve3d_point3d_distance_far() {
         let curve = create_test_curve::<f64>();
         let collider = NurbsCurveCollider::new(curve);
-        let point = Point3D::new(0.1, 0.0, 0.0); // 曲線の近く
-        let distance = collider.distance_to(&point);
-
-        assert!(distance < 0.2); // 適度に近い
-        assert!(distance > 0.0); // 完全には一致しない
-    }
-
-    #[test]
-    fn test_point_far_from_curve() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-        let point = Point3D::new(10.0, 10.0, 10.0); // 曲線から遠い
+        let point = CorePoint3D::new(10.0, 10.0, 10.0);
         let distance = collider.distance_to(&point);
 
         assert!(distance > 10.0);
     }
 
     #[test]
-    fn test_symmetry() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
+    fn test_nurbssurface3d_point3d_distance_zero_on_surface() {
+        let surface = create_test_surface::<f64>();
         let point = Point3D::new(0.5, 0.5, 0.0);
-
-        // Newtype パターンのため一方向のみテスト
-        let distance = collider.distance_to(&point);
-        assert!(distance >= 0.0); // 距離は非負
+        let d = nurbssurface3d_point3d_distance(&surface, &point);
+        assert!(d <= 1e-6);
     }
 
     #[test]
-    fn test_newtype_conversions() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve.clone());
-
-        // inner() で参照を取得
-        let _inner_ref = collider.inner();
-
-        // into_inner() で元の NurbsCurve3D を取得
-        let recovered_curve = collider.into_inner();
-
-        // 元の曲線と同じパラメータ範囲を持つことを確認
-        assert_eq!(recovered_curve.parameter_domain(), curve.parameter_domain());
-    }
-
-    // LineSegment3D tests
-
-    #[test]
-    fn test_line_segment_intersecting() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の始点を通る線分
-        let segment =
-            LineSegment3D::new(Point3D::new(-0.1, 0.0, 0.0), Point3D::new(0.1, 0.0, 0.0)).unwrap();
-
-        let tolerance = 0.1;
-        assert!(collider.intersects(&segment, tolerance));
-    }
-
-    #[test]
-    fn test_line_segment_near() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の近くの線分
-        let segment =
-            LineSegment3D::new(Point3D::new(0.5, 0.5, 0.5), Point3D::new(1.5, 0.5, 0.5)).unwrap();
-
-        let distance = collider.distance_to(&segment);
-        assert!(distance > 0.0);
-        assert!(distance < 1.0);
-    }
-
-    // Ray3D tests
-
-    #[test]
-    fn test_ray_intersecting() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の始点を通るRay
-        let ray = Ray3D::new(
-            Point3D::new(-1.0, 0.0, 0.0),
-            geo_core::Vector3D::new(1.0, 0.0, 0.0),
-        )
-        .unwrap();
-
-        let tolerance = 0.1;
-        assert!(collider.intersects(&ray, tolerance));
-    }
-
-    // InfiniteLine3D tests
-
-    #[test]
-    fn test_infinite_line_intersecting() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の始点を通る無限直線
-        let line = InfiniteLine3D::from_two_points(
-            Point3D::new(-1.0, 0.0, 0.0),
-            Point3D::new(1.0, 0.0, 0.0),
-        )
-        .unwrap();
-
-        let tolerance = 0.1;
-        assert!(collider.intersects(&line, tolerance));
-    }
-
-    // Circle3D tests
-
-    #[test]
-    fn test_circle_near() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の近くの円
-        use crate::Direction3D;
-        let circle = Circle3D::new(
-            Point3D::new(0.5, 0.5, 0.0),
-            Direction3D::from_vector(geo_core::Vector3D::new(0.0, 0.0, 1.0)).unwrap(),
-            0.1,
-        )
-        .unwrap();
-
-        let distance = collider.distance_to(&circle);
-        assert!(distance >= 0.0);
-    }
-
-    // Plane3D tests
-
-    #[test]
-    fn test_plane_intersecting() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線が乗っているxy平面
+    fn test_nurbssurface3d_plane3d_distance_zero_for_xy_plane() {
+        let surface = create_test_surface::<f64>();
         let plane = Plane3D::xy_plane(0.0);
-
-        let tolerance = 1e-6;
-        assert!(collider.intersects(&plane, tolerance));
+        let d = nurbssurface3d_plane3d_distance(&surface, &plane);
+        assert!(d <= 1e-6);
     }
 
     #[test]
-    fn test_plane_distance() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線から離れた平面（z = 1.0）
-        let plane = Plane3D::xy_plane(1.0);
-
-        let distance = collider.distance_to(&plane);
-        assert!((distance - 1.0).abs() < 0.1); // 約1.0の距離
-    }
-
-    // SphericalSolid3D tests
-
-    #[test]
-    fn test_spherical_solid_intersecting() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の始点を含む球
-        use geo_core::Vector3D;
-        let sphere = SphericalSolid3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
+    fn test_nurbssurface3d_ray3d_distance_zero_for_vertical_hit() {
+        let surface = create_test_surface::<f64>();
+        let ray = Ray3D::new(
+            CorePoint3D::new(0.5, 0.5, -1.0),
             Vector3D::new(0.0, 0.0, 1.0),
-            Vector3D::new(1.0, 0.0, 0.0),
-            0.5,
         )
         .unwrap();
-
-        let tolerance = 1e-6;
-        assert!(collider.intersects(&sphere, tolerance));
-    }
-
-    #[test]
-    fn test_spherical_solid_separate() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線から離れた球
-        use geo_core::Vector3D;
-        let sphere = SphericalSolid3D::new(
-            Point3D::new(10.0, 10.0, 10.0),
-            Vector3D::new(0.0, 0.0, 1.0),
-            Vector3D::new(1.0, 0.0, 0.0),
-            0.5,
-        )
-        .unwrap();
-
-        let distance = collider.distance_to(&sphere);
-        assert!(distance > 10.0);
-    }
-
-    // EllipsoidalSolid3D tests
-
-    #[test]
-    fn test_ellipsoidal_solid_intersecting() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の始点を含む楕円体
-        use geo_core::Vector3D;
-        let ellipsoid = EllipsoidalSolid3D::new(
-            Point3D::new(0.0, 0.0, 0.0),
-            Vector3D::new(0.0, 0.0, 1.0),
-            Vector3D::new(1.0, 0.0, 0.0),
-            0.5,
-            0.5,
-            0.5,
-        )
-        .unwrap();
-
-        let tolerance = 1e-6;
-        assert!(collider.intersects(&ellipsoid, tolerance));
-    }
-
-    // CylindricalSolid3D tests
-
-    #[test]
-    fn test_cylindrical_solid_near() {
-        let curve = create_test_curve::<f64>();
-        let collider = NurbsCurveCollider::new(curve);
-
-        // 曲線の近くの円柱
-        use geo_core::Vector3D;
-        let cylinder = CylindricalSolid3D::new(
-            Point3D::new(0.5, 0.5, -1.0),
-            Vector3D::new(0.0, 0.0, 1.0),
-            Vector3D::new(1.0, 0.0, 0.0),
-            0.1,
-            2.0,
-        )
-        .unwrap();
-
-        let distance = collider.distance_to(&cylinder);
-        assert!(distance >= 0.0);
+        let d = nurbssurface3d_ray3d_distance(&surface, &ray);
+        assert!(d <= 1e-6);
     }
 }
