+++ model/geo_algorithms/src/distance/primitive_3d.rs
@@ -5,8 +5,8 @@
 //!
 //! 命名規則: `{shape_a}_{shape_b}_distance`
 
-use crate::{Circle3D, InfiniteLine3D, LineSegment3D, NurbsCurve3D, Plane3D, Point3D, Ray3D};
-use geo_contracts::{DistanceConvergenceError, Scalar};
+use crate::{Circle3D, InfiniteLine3D, LineSegment3D, Plane3D, Point3D, Ray3D};
+use geo_contracts::Scalar;
 
 /// LineSegment3D-点 間の最短距離（端点クランプあり）
 pub fn line_segment3d_point3d_distance<T: Scalar>(
@@ -78,75 +78,18 @@ pub fn point3d_circle3d_distance<T: Scalar>(point: &Point3D<T>, circle: &Circle3
     circle.distance_to_point_3d(*point)
 }
 
-/// NurbsCurve3D-点 間の最短距離（フォールバック付き）
-pub fn nurbscurve3d_point3d_distance<T: Scalar>(curve: &NurbsCurve3D<T>, point: &Point3D<T>) -> T {
-    match nurbscurve3d_point3d_try_distance(curve, point, 64) {
-        Ok(distance) => distance,
-        Err(_) => crate::collision::nurbscurve3d_point3d_distance(curve, point),
-    }
-}
-
-/// 逆向きラッパー: point-nurbscurve
-pub fn point3d_nurbscurve3d_distance<T: Scalar>(point: &Point3D<T>, curve: &NurbsCurve3D<T>) -> T {
-    nurbscurve3d_point3d_distance(curve, point)
-}
-
-/// NurbsCurve3D-点 間の最短距離（fallible）
-pub fn nurbscurve3d_point3d_try_distance<T: Scalar>(
-    curve: &NurbsCurve3D<T>,
-    point: &Point3D<T>,
-    max_samples: usize,
-) -> Result<T, DistanceConvergenceError<T>> {
-    if max_samples == 0 {
-        return Err(DistanceConvergenceError::InvalidInitialization);
-    }
-
-    if max_samples == 1 {
-        return Err(DistanceConvergenceError::NotConverged {
-            iterations: 1,
-            residual: T::INFINITY,
-        });
-    }
-
-    let (u_min, u_max) = curve.parameter_domain();
-    let delta_u = (u_max - u_min) / T::from_usize(max_samples);
-
-    let mut min_distance = T::INFINITY;
-    for i in 0..=max_samples {
-        let u = u_min + delta_u * T::from_usize(i);
-        let curve_vec = curve.evaluate_at(u);
-        let curve_point = Point3D::new(curve_vec.x(), curve_vec.y(), curve_vec.z());
-        min_distance = min_distance.min(curve_point.distance_to(point));
-    }
-
-    if min_distance == T::INFINITY {
-        return Err(DistanceConvergenceError::NumericalFailure);
-    }
-
-    Ok(min_distance)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use geo_contracts::NurbsCurve3DConstructor;
+    use geo_contracts::ToleranceSettings;
 
-    fn create_test_curve() -> NurbsCurve3D<f64> {
-        let control_points = vec![
-            (0.0, 0.0, 0.0),
-            (1.0, 0.0, 0.0),
-            (1.0, 1.0, 0.0),
-            (0.0, 1.0, 0.0),
-        ];
-        let knots = vec![0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0];
-        let weights = Some(vec![1.0, 1.0, 1.0, 1.0]);
-
-        <NurbsCurve3D<f64> as NurbsCurve3DConstructor<f64>>::new(3, knots, control_points, weights)
-            .unwrap()
+    fn standard_distance_tol() -> f64 {
+        ToleranceSettings::<f64>::standard().distance_tolerance
     }
 
     #[test]
     fn infinite_line3d_distance_parallel_lines() {
+        let tol = standard_distance_tol();
         let line_a = InfiniteLine3D::from_two_points(
             Point3D::new(0.0, 0.0, 0.0),
             Point3D::new(1.0, 0.0, 0.0),
@@ -159,11 +102,12 @@ mod tests {
         .unwrap();
 
         let d = infinite_line3d_infinite_line3d_distance(&line_a, &line_b);
-        assert!((d - 2.0).abs() < 1e-10);
+        assert!((d - 2.0).abs() < tol);
     }
 
     #[test]
     fn infinite_line3d_distance_intersecting_lines() {
+        let tol = standard_distance_tol();
         let line_a = InfiniteLine3D::from_two_points(
             Point3D::new(-1.0, 0.0, 0.0),
             Point3D::new(1.0, 0.0, 0.0),
@@ -176,40 +120,6 @@ mod tests {
         .unwrap();
 
         let d = infinite_line3d_infinite_line3d_distance(&line_a, &line_b);
-        assert!(d.abs() < 1e-10);
-    }
-
-    #[test]
-    fn nurbscurve_try_distance_rejects_zero_samples() {
-        let curve = create_test_curve();
-        let point = Point3D::new(0.2, 0.2, 0.0);
-
-        let result = nurbscurve3d_point3d_try_distance(&curve, &point, 0);
-        assert!(matches!(
-            result,
-            Err(DistanceConvergenceError::InvalidInitialization)
-        ));
-    }
-
-    #[test]
-    fn nurbscurve_try_distance_reports_not_converged_for_one_sample() {
-        let curve = create_test_curve();
-        let point = Point3D::new(0.2, 0.2, 0.0);
-
-        let result = nurbscurve3d_point3d_try_distance(&curve, &point, 1);
-        assert!(matches!(
-            result,
-            Err(DistanceConvergenceError::NotConverged { iterations: 1, .. })
-        ));
-    }
-
-    #[test]
-    fn point_to_nurbscurve_wrapper_is_symmetric() {
-        let curve = create_test_curve();
-        let point = Point3D::new(0.2, 0.2, 0.0);
-
-        let a_to_b = nurbscurve3d_point3d_distance(&curve, &point);
-        let b_to_a = point3d_nurbscurve3d_distance(&point, &curve);
-        assert!((a_to_b - b_to_a).abs() < 1e-10);
+        assert!(d.abs() < tol);
     }
 }
