+++ model/geo_algorithms/tests/coverage_matrix_engine.rs
@@ -179,11 +179,13 @@ fn known_entrypoints() -> HashSet<String> {
     let sources = [
         include_str!("../src/collision/primitive_2d.rs"),
         include_str!("../src/collision/primitive_3d.rs"),
-        include_str!("../src/collision/primitive_nurbs.rs"),
+        include_str!("../src/collision/nurbs_3d.rs"),
         include_str!("../src/intersection/primitive_2d.rs"),
         include_str!("../src/intersection/primitive_3d.rs"),
+        include_str!("../src/intersection/nurbs_3d.rs"),
         include_str!("../src/distance/primitive_2d.rs"),
         include_str!("../src/distance/primitive_3d.rs"),
+        include_str!("../src/distance/nurbs_3d.rs"),
     ];
 
     let mut set = HashSet::new();
@@ -1220,7 +1222,7 @@ fn valid_seed_entries() -> Vec<MatrixEntry> {
             entrypoint_b_to_a: Some("circle3d_point3d_distance"),
         },
         MatrixEntry {
-            id: "3d:nurbscurve-point:distance",
+            id: "3d:nurbs_3d-point:distance",
             dimension: Dimension::D3,
             operation: Operation::Distance,
             shape_a: "NurbsCurve3D",
@@ -1232,7 +1234,7 @@ fn valid_seed_entries() -> Vec<MatrixEntry> {
             entrypoint_b_to_a: Some("point3d_nurbscurve3d_distance"),
         },
         MatrixEntry {
-            id: "3d:point-nurbscurve:distance",
+            id: "3d:point-nurbs_3d:distance",
             dimension: Dimension::D3,
             operation: Operation::Distance,
             shape_a: "Point3D",
@@ -1242,6 +1244,198 @@ fn valid_seed_entries() -> Vec<MatrixEntry> {
             cardinality: None,
             entrypoint_a_to_b: Some("point3d_nurbscurve3d_distance"),
             entrypoint_b_to_a: Some("nurbscurve3d_point3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-linesegment:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "LineSegment3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_linesegment3d_distance"),
+            entrypoint_b_to_a: Some("linesegment3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:linesegment-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "LineSegment3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("linesegment3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_linesegment3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-ray:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "Ray3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_ray3d_distance"),
+            entrypoint_b_to_a: Some("ray3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:ray-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "Ray3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("ray3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_ray3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-infinite_line:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "InfiniteLine3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_infinite_line3d_distance"),
+            entrypoint_b_to_a: Some("infinite_line3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:infinite_line-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "InfiniteLine3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("infinite_line3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_infinite_line3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-circle:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "Circle3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_circle3d_distance"),
+            entrypoint_b_to_a: Some("circle3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:circle-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "Circle3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("circle3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_circle3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-plane:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "Plane3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_plane3d_distance"),
+            entrypoint_b_to_a: Some("plane3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:plane-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "Plane3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("plane3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_plane3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-spherical_solid:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "SphericalSolid3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_spherical_solid3d_distance"),
+            entrypoint_b_to_a: Some("spherical_solid3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:spherical_solid-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "SphericalSolid3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("spherical_solid3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_spherical_solid3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-ellipsoidal_solid:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "EllipsoidalSolid3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_ellipsoidal_solid3d_distance"),
+            entrypoint_b_to_a: Some("ellipsoidal_solid3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:ellipsoidal_solid-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "EllipsoidalSolid3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("ellipsoidal_solid3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_ellipsoidal_solid3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:nurbs_3d-cylindrical_solid:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "NurbsCurve3D",
+            shape_b: "CylindricalSolid3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("nurbscurve3d_cylindrical_solid3d_distance"),
+            entrypoint_b_to_a: Some("cylindrical_solid3d_nurbscurve3d_distance"),
+        },
+        MatrixEntry {
+            id: "3d:cylindrical_solid-nurbs_3d:distance",
+            dimension: Dimension::D3,
+            operation: Operation::Distance,
+            shape_a: "CylindricalSolid3D",
+            shape_b: "NurbsCurve3D",
+            required: true,
+            symmetric: true,
+            cardinality: None,
+            entrypoint_a_to_b: Some("cylindrical_solid3d_nurbscurve3d_distance"),
+            entrypoint_b_to_a: Some("nurbscurve3d_cylindrical_solid3d_distance"),
         },
     ]
 }
